refactor: torna WithSearch um módulo reutilizável em qualquer indicador e perfil
