+++ src/features/acf/common/frontend/WithSearch/context.ts
@@ -1,0 +1,10 @@
+"use client";
+import { createContext } from "react";
+
+export type SearchModel = {
+    searchString: string;
+};
+
+export const SearchContext = createContext<SearchModel>({
+    searchString: "",
+});
